requirements.txt
torch
transformers
pandas
numpy
tqdm
biopython
scikit-learn